requirements.txt
flask>=1.0.0
ruamel.yaml
peewee
pandas
visdom
dataclasses
peewee
